chore: big refactor of keygen.rs
