+++ core/threshold/src/execution/endpoints/keygen.rs
@@ -1,67 +1,66 @@
-use crate::algebra::base_ring::{Z128, Z64};
-use crate::algebra::structure_traits::Zero;
-use crate::execution::online::preprocessing::{DKGPreprocessing, RandomPreprocessing};
-use crate::execution::tfhe_internals::compression_decompression_key::{
-    CompressionPrivateKeyShares, SnsCompressionPrivateKeyShares,
-};
-use crate::execution::tfhe_internals::lwe_bootstrap_key::par_decompress_into_lwe_bootstrap_key_generated_from_xof;
-use crate::execution::tfhe_internals::lwe_ciphertext::{
-    self, encrypt_lwe_ciphertext_list, LweCiphertextShare,
-};
-use crate::execution::tfhe_internals::lwe_key::LweCompactPublicKeyShare;
-use crate::execution::tfhe_internals::lwe_keyswitch_key::LweKeySwitchKeyShare;
-use crate::execution::tfhe_internals::lwe_packing_keyswitch_key_generation::allocate_and_generate_lwe_packing_keyswitch_key;
-use crate::execution::tfhe_internals::parameters::{
-    BKParams, DKGParams, DistributedCompressionParameters, DistributedSnsCompressionParameters,
-    EncryptionType, KSKParams, MSNRKConfiguration, MSNRKParams, NoiseInfo,
-};
-use crate::execution::tfhe_internals::private_keysets::{GenericPrivateKeySet, PrivateKeySet};
-use crate::execution::tfhe_internals::public_keysets::{
-    CompressedFhePubKeySet, FhePubKeySet, RawCompressedPubKeySet, RawPubKeySet,
-};
-use crate::hashing::DomainSep;
 use crate::{
     algebra::{
+        base_ring::{Z128, Z64},
         galois_rings::common::ResiduePoly,
         structure_traits::{BaseRing, ErrorCorrect, Ring},
     },
     error::error_handler::anyhow_error_and_log,
     execution::{
-        online::triple::open_list,
+        online::{
+            preprocessing::{DKGPreprocessing, RandomPreprocessing},
+            triple::open_list,
+        },
         runtime::session::BaseSessionHandles,
         tfhe_internals::{
+            compression_decompression_key::CompressionPrivateKeyShares,
+            compression_decompression_key_generation::{
+                distributed_keygen_compressed_compression_material,
+                distributed_keygen_compression_material, generate_compressed_decompression_keys,
+                generate_decompression_keys,
+            },
             glwe_key::GlweSecretKeyShare,
-            lwe_bootstrap_key_generation::allocate_and_generate_lwe_bootstrap_key,
-            lwe_key::{allocate_and_generate_new_lwe_compact_public_key, LweSecretKeyShare},
-            lwe_keyswitch_key_generation::allocate_and_generate_new_lwe_keyswitch_key,
-            lwe_packing_keyswitch_key::LwePackingKeyswitchKeyShares,
+            lwe_bootstrap_key::par_decompress_into_lwe_bootstrap_key_generated_from_xof,
+            lwe_bootstrap_key_generation::{
+                generate_bootstrap_key, generate_compressed_bootstrap_key,
+            },
+            lwe_key::{
+                generate_lwe_private_compressed_public_key_pair,
+                generate_lwe_private_public_key_pair, LweSecretKeyShare,
+            },
+            lwe_keyswitch_key_generation::{
+                generate_compressed_key_switch_key, generate_key_switch_key,
+            },
+            modulus_switch_noise_reduction_key_generation::{
+                generate_compressed_mod_switch_noise_reduction_key,
+                generate_mod_switch_noise_reduction_key,
+            },
+            parameters::{DKGParams, MSNRKConfiguration},
+            private_keysets::{GenericPrivateKeySet, PrivateKeySet},
+            public_keysets::{
+                CompressedFhePubKeySet, FhePubKeySet, RawCompressedPubKeySet, RawPubKeySet,
+            },
             randomness::MPCEncryptionRandomGenerator,
+            sns_compression_key::SnsCompressionPrivateKeyShares,
+            sns_compression_key_generation::{
+                generate_compressed_sns_compression_keys, generate_sns_compression_keys,
+            },
         },
     },
+    hashing::DomainSep,
 };
-use itertools::Itertools;
 use num_integer::div_ceil;
-use tfhe::core_crypto::commons::traits::UnsignedInteger;
-use tfhe::core_crypto::prelude::{
-    SeededLweBootstrapKey, SeededLweCompactPublicKey, SeededLweKeyswitchKey,
+use tfhe::{
+    core_crypto::entities::LweBootstrapKey,
+    shortint::{
+        list_compression::{
+            CompressedDecompressionKey, DecompressionKey, NoiseSquashingCompressionKey,
+        },
+        parameters::LweCiphertextCount,
+        server_key::{CompressedModulusSwitchConfiguration, ModulusSwitchConfiguration},
+        ClassicPBSParameters,
+    },
 };
-use tfhe::core_crypto::{
-    algorithms::par_convert_standard_lwe_bootstrap_key_to_fourier,
-    entities::{FourierLweBootstrapKey, LweBootstrapKey, LweCompactPublicKey, LweKeyswitchKey},
-    prelude::ByteRandomGenerator,
-};
-use tfhe::shortint::list_compression::{
-    CompressedCompressionKey, CompressedDecompressionKey, CompressedNoiseSquashingCompressionKey,
-    CompressionKey, DecompressionKey, NoiseSquashingCompressionKey,
-};
-use tfhe::shortint::parameters::LweCiphertextCount;
-use tfhe::shortint::server_key::{
-    CompressedModulusSwitchConfiguration, CompressedModulusSwitchNoiseReductionKey,
-    ModulusSwitchConfiguration, ModulusSwitchNoiseReductionKey, ShortintBootstrappingKey,
-};
-use tfhe::shortint::ClassicPBSParameters;
-use tfhe_csprng::generators::SoftwareRandomGenerator;
-use tfhe_csprng::seeders::XofSeed;
+use tfhe_csprng::{generators::SoftwareRandomGenerator, seeders::XofSeed};
 use tracing::instrument;
 
 pub(crate) const DSEP_KG: DomainSep = *b"TFHE_GEN";
@@ -87,961 +86,6 @@ async fn sample_seed<
         .flat_map(Z::to_byte_vec)
         .take((sec as usize) >> 3)
         .fold(0_u128, |acc, x| (acc << 8) + (x as u128)))
-}
-
-///Generates the lwe private key share and associated public key
-fn generate_lwe_key_shares<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    params: &DKGParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<(
-    LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    LweCompactPublicKeyShare<Z, EXTENSION_DEGREE>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let params = params.get_params_basics_handle();
-    let my_role = session.my_role();
-    //Init the shared LWE secret key
-    tracing::info!("(Party {my_role}) Generating LWE Secret key...Start");
-    let lwe_secret_key_share =
-        LweSecretKeyShare::new_from_preprocessing(params.lwe_hat_dimension(), preprocessing)?;
-    tracing::info!("(Party {my_role}) Generating corresponding public key...Start");
-    let NoiseInfo { amount, bound } = params.num_needed_noise_pk();
-    let vec_tuniform_noise = preprocessing
-        .next_noise_vec(amount, bound)?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    //and fill the noise generator with noise generated above
-    mpc_encryption_rng.fill_noise(vec_tuniform_noise);
-
-    //Then actually generate the public key
-    let lwe_public_key_shared =
-        allocate_and_generate_new_lwe_compact_public_key(&lwe_secret_key_share, mpc_encryption_rng);
-
-    Ok((lwe_secret_key_share, lwe_public_key_shared))
-}
-
-///Generates the lwe private key share and associated public key
-#[instrument(name="Gen Lwe keys",skip( mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_lwe_private_public_key_pair<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    params: &DKGParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<(
-    LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    LweCompactPublicKey<Vec<u64>>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let (lwe_secret_key_share, lwe_public_key_shared) =
-        generate_lwe_key_shares(params, mpc_encryption_rng, session, preprocessing)?;
-
-    //Open the public key and cast it to TFHE-RS type
-    Ok((
-        lwe_secret_key_share,
-        lwe_public_key_shared.open_to_tfhers_type(session).await?,
-    ))
-}
-
-///Generates the lwe private key share and associated public key
-#[instrument(name="Gen compressed Lwe keys",skip( mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_lwe_private_compressed_public_key_pair<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    params: &DKGParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<(
-    LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    SeededLweCompactPublicKey<Vec<u64>>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let (lwe_secret_key_share, lwe_public_key_shared) =
-        generate_lwe_key_shares(params, mpc_encryption_rng, session, preprocessing)?;
-
-    //Open the public key and cast it to TFHE-RS type
-    Ok((
-        lwe_secret_key_share,
-        lwe_public_key_shared
-            .open_to_tfhers_seeded_type(seed, session)
-            .await?,
-    ))
-}
-
-/// Generate the modulus switching noise reduction key from the small LWE key.
-/// This key is essentially encryptions of zeros, and it's used as a part of
-/// the bootstrap algorithm if it exists, right before modulus switching.
-#[instrument(name="Gen MSNRK",skip(input_lwe_sk, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_mod_switch_noise_reduction_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &MSNRKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<ModulusSwitchNoiseReductionKey<u64>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let output = generate_encrypted_zeros(
-        input_lwe_sk,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )?;
-
-    let opened_ciphertext_list = lwe_ciphertext::open_to_tfhers_type(output, session).await?;
-
-    Ok(ModulusSwitchNoiseReductionKey {
-        modulus_switch_zeros: opened_ciphertext_list,
-        ms_bound: params.params.ms_bound,
-        ms_r_sigma_factor: params.params.ms_r_sigma_factor,
-        ms_input_variance: params.params.ms_input_variance,
-    })
-}
-
-#[instrument(name="Gen Compressed MSNRK",skip(input_lwe_sk, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_mod_switch_noise_reduction_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &MSNRKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<CompressedModulusSwitchNoiseReductionKey<u64>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let output = generate_encrypted_zeros(
-        input_lwe_sk,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )?;
-
-    let opened_ciphertext_list =
-        lwe_ciphertext::open_to_tfhers_seeded_type(output, seed, session).await?;
-
-    Ok(CompressedModulusSwitchNoiseReductionKey {
-        modulus_switch_zeros: opened_ciphertext_list,
-        ms_bound: params.params.ms_bound,
-        ms_r_sigma_factor: params.params.ms_r_sigma_factor,
-        ms_input_variance: params.params.ms_input_variance,
-    })
-}
-
-fn generate_encrypted_zeros<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &MSNRKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<Vec<LweCiphertextShare<Z, EXTENSION_DEGREE>>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let my_role = session.my_role();
-    tracing::info!("(Party {my_role}) Generating MSNRK...Start");
-    let vec_tuniform_noise = preprocessing
-        .next_noise_vec(params.num_needed_noise, params.noise_bound)?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    mpc_encryption_rng.fill_noise(vec_tuniform_noise);
-
-    let zeros_count = params.params.modulus_switch_zeros_count.0;
-    let lwe_size = input_lwe_sk.lwe_dimension().to_lwe_size();
-    let mut output = vec![LweCiphertextShare::new(lwe_size); zeros_count];
-    let encoded = vec![ResiduePoly::<Z, EXTENSION_DEGREE>::ZERO; zeros_count];
-    encrypt_lwe_ciphertext_list(input_lwe_sk, &mut output, &encoded, mpc_encryption_rng)?;
-
-    Ok(output)
-}
-
-/// Generate KSK shares using MPC encryption
-fn generate_ksk_share<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    output_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &KSKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<LweKeySwitchKeyShare<Z, EXTENSION_DEGREE>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let my_role = session.my_role();
-    tracing::info!("(Party {my_role}) Generating KSK...Start");
-    let vec_tuniform_noise = preprocessing
-        .next_noise_vec(params.num_needed_noise, params.noise_bound)?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    mpc_encryption_rng.fill_noise(vec_tuniform_noise);
-
-    //Then compute the KSK
-    allocate_and_generate_new_lwe_keyswitch_key(
-        input_lwe_sk,
-        output_lwe_sk,
-        params.decomposition_base_log,
-        params.decomposition_level_count,
-        mpc_encryption_rng,
-    )
-}
-
-///Generate the Key Switch Key from a Glwe key given in Lwe format,
-///and an actual Lwe key
-#[instrument(name="Gen KSK",skip(input_lwe_sk, output_lwe_sk, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_key_switch_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    output_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &KSKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<LweKeyswitchKey<Vec<u64>>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let ksk_share = generate_ksk_share(
-        input_lwe_sk,
-        output_lwe_sk,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )?;
-
-    //Open the KSK and cast it to TFHE-RS type
-    ksk_share.open_to_tfhers_type(session).await
-}
-
-///Generate the Key Switch Key from a Glwe key given in Lwe format,
-///and an actual Lwe key
-#[instrument(name="Gen compressed KSK",skip(input_lwe_sk, output_lwe_sk, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_key_switch_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    input_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    output_lwe_sk: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &KSKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<SeededLweKeyswitchKey<Vec<u64>>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let ksk_share = generate_ksk_share(
-        input_lwe_sk,
-        output_lwe_sk,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )?;
-
-    //Open the KSK and cast it to TFHE-RS seeded type
-    ksk_share.open_to_tfhers_seeded_type(seed, session).await
-}
-
-///Helper function to generate bootstrap key share
-async fn generate_bootstrap_key_share<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    lwe_secret_key_share: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &BKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<
-    crate::execution::tfhe_internals::lwe_bootstrap_key::LweBootstrapKeyShare<Z, EXTENSION_DEGREE>,
->
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let my_role = session.my_role();
-    //First sample the noise
-    let vec_tuniform_noise = preprocessing
-        .next_noise_vec(params.num_needed_noise, params.noise_bound)?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    mpc_encryption_rng.fill_noise(vec_tuniform_noise);
-
-    tracing::info!("(Party {my_role}) Generating BK for {:?} ...Start", params);
-
-    let bk_share = allocate_and_generate_lwe_bootstrap_key(
-        lwe_secret_key_share,
-        glwe_secret_key_share,
-        params.decomposition_base_log,
-        params.decomposition_level_count,
-        mpc_encryption_rng,
-        params.enc_type,
-        session,
-        preprocessing,
-    )
-    .await?;
-
-    tracing::info!("(Party {my_role}) Generating BK {:?} ...Done", params);
-    Ok(bk_share)
-}
-
-///Generates a Bootstrapping Key given a Glwe key in Glwe format
-///, a Lwe key and the params for the BK generation
-#[instrument(name="Gen BK", skip(glwe_secret_key_share, lwe_secret_key_share, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_bootstrap_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    Scalar: UnsignedInteger,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    lwe_secret_key_share: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: BKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<LweBootstrapKey<Vec<Scalar>>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let my_role = session.my_role();
-    let bk_share = generate_bootstrap_key_share(
-        glwe_secret_key_share,
-        lwe_secret_key_share,
-        &params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )
-    .await?;
-
-    tracing::info!("(Party {my_role}) Opening BK {:?} ...Start", params);
-    //Open the bk and cast it to TFHE-rs type
-    let bk = bk_share.open_to_tfhers_type::<Scalar, _>(session).await?;
-    tracing::info!("(Party {my_role}) Opening BK {:?} ...Done", params);
-    Ok(bk)
-}
-
-///Generates a compressed Bootstrapping Key given a Glwe key in Glwe format
-///, a Lwe key and the params for the BK generation
-#[instrument(name="Gen compressed BK", skip(glwe_secret_key_share, lwe_secret_key_share, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_bootstrap_key<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    Scalar: UnsignedInteger,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    lwe_secret_key_share: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: BKParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<SeededLweBootstrapKey<Vec<Scalar>>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let my_role = session.my_role();
-    let bk_share = generate_bootstrap_key_share(
-        glwe_secret_key_share,
-        lwe_secret_key_share,
-        &params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )
-    .await?;
-
-    tracing::info!("(Party {my_role}) Opening BK {:?} ...Start", params);
-    //Open the bk and cast it to TFHE-rs type
-    let bk = bk_share
-        .open_to_tfhers_seeded_type::<Scalar, _>(seed, session)
-        .await?;
-    tracing::info!("(Party {my_role}) Opening BK {:?} ...Done", params);
-    Ok(bk)
-}
-
-#[instrument(name="Gen Decompression Key", skip(private_glwe_compute_key, private_compression_key, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_decompression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    private_glwe_compute_key: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_compression_key: &CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    params: DistributedCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<DecompressionKey>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let blind_rotate_key: LweBootstrapKey<Vec<u64>> = generate_bootstrap_key(
-        private_glwe_compute_key,
-        &private_compression_key.clone().into_lwe_secret_key(),
-        params.bk_params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )
-    .await?;
-
-    // Creation of the bootstrapping key in the Fourier domain
-    let mut fourier_bsk = FourierLweBootstrapKey::new(
-        blind_rotate_key.input_lwe_dimension(),
-        blind_rotate_key.glwe_size(),
-        blind_rotate_key.polynomial_size(),
-        blind_rotate_key.decomposition_base_log(),
-        blind_rotate_key.decomposition_level_count(),
-    );
-
-    // Conversion to fourier domain
-    par_convert_standard_lwe_bootstrap_key_to_fourier(&blind_rotate_key, &mut fourier_bsk);
-
-    // TODO implement `modulus_switch_noise_reduction_key` keygen
-    let blind_rotate_key = ShortintBootstrappingKey::Classic {
-        bsk: fourier_bsk,
-        // NOTE: Not sure if it should be standard or CenteredMean
-        modulus_switch_noise_reduction_key: ModulusSwitchConfiguration::Standard,
-    };
-
-    Ok(DecompressionKey {
-        blind_rotate_key,
-        lwe_per_glwe: params.raw_compression_parameters.lwe_per_glwe,
-    })
-}
-
-#[instrument(name="Gen compressed Decompression Key", skip(private_glwe_compute_key, private_compression_key, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_decompression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    private_glwe_compute_key: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_compression_key: &CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    params: DistributedCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<CompressedDecompressionKey>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let blind_rotate_key: SeededLweBootstrapKey<Vec<u64>> = generate_compressed_bootstrap_key(
-        private_glwe_compute_key,
-        &private_compression_key.clone().into_lwe_secret_key(),
-        params.bk_params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-        seed,
-    )
-    .await?;
-
-    Ok(CompressedDecompressionKey {
-        blind_rotate_key,
-        lwe_per_glwe: params.raw_compression_parameters.lwe_per_glwe,
-    })
-}
-
-fn generate_sns_compression_key_shares<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share_sns_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: &DistributedSnsCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    preprocessing: &mut P,
-) -> anyhow::Result<(
-    SnsCompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    LwePackingKeyswitchKeyShares<Z, EXTENSION_DEGREE>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let private_sns_compression_key_shares =
-        SnsCompressionPrivateKeyShares::new_from_preprocessing(
-            params.raw_compression_parameters,
-            preprocessing,
-        )
-        .inspect_err(|e| {
-            tracing::error!("failed to generate private sns compression shares: {e}")
-        })?;
-
-    let noise_vec = preprocessing
-        .next_noise_vec(params.ksk_num_noise, params.ksk_noisebound)
-        .inspect_err(|e| {
-            tracing::error!("failed to get noise vec for sns compression shares: {e}")
-        })?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    mpc_encryption_rng.fill_noise(noise_vec);
-
-    let packing_key_switching_key_shares = allocate_and_generate_lwe_packing_keyswitch_key(
-        glwe_secret_key_share_sns_as_lwe,
-        &private_sns_compression_key_shares.post_packing_ks_key,
-        params.raw_compression_parameters.packing_ks_base_log,
-        params.raw_compression_parameters.packing_ks_level,
-        EncryptionType::Bits128,
-        mpc_encryption_rng,
-    );
-
-    Ok((
-        private_sns_compression_key_shares,
-        packing_key_switching_key_shares,
-    ))
-}
-
-#[instrument(name="Gen Sns Compression Key", skip(glwe_secret_key_share_sns_as_lwe, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_sns_compression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share_sns_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: DistributedSnsCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<(
-    SnsCompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    NoiseSquashingCompressionKey,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let (private_sns_compression_key_shares, packing_key_switching_key_shares) =
-        generate_sns_compression_key_shares(
-            glwe_secret_key_share_sns_as_lwe,
-            &params,
-            mpc_encryption_rng,
-            preprocessing,
-        )?;
-
-    let packing_key_switching_key = packing_key_switching_key_shares
-        .open_to_tfhers_type::<u128, _>(session)
-        .await
-        .inspect_err(|e| tracing::error!("failed to open tfhers type u128: {e}"))?;
-
-    let compression_key = NoiseSquashingCompressionKey::from_raw_parts(
-        packing_key_switching_key,
-        params.raw_compression_parameters.lwe_per_glwe,
-    );
-    Ok((private_sns_compression_key_shares, compression_key))
-}
-
-#[instrument(name="Gen compressed Sns Compression Key", skip(glwe_secret_key_share_sns_as_lwe, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_sns_compression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    glwe_secret_key_share_sns_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    params: DistributedSnsCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<(
-    SnsCompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    CompressedNoiseSquashingCompressionKey,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let (private_sns_compression_key_shares, packing_key_switching_key_shares) =
-        generate_sns_compression_key_shares(
-            glwe_secret_key_share_sns_as_lwe,
-            &params,
-            mpc_encryption_rng,
-            preprocessing,
-        )?;
-
-    let packing_key_switching_key = packing_key_switching_key_shares
-        .open_to_tfhers_seeded_type::<u128, _>(seed, session)
-        .await
-        .inspect_err(|e| tracing::error!("failed to open tfhers type u128: {e}"))?;
-
-    let compression_key = CompressedNoiseSquashingCompressionKey::from_raw_parts(
-        packing_key_switching_key,
-        params.raw_compression_parameters.lwe_per_glwe,
-    );
-
-    Ok((private_sns_compression_key_shares, compression_key))
-}
-
-/// Helper function to generate packing key switching key shares
-/// for use in Compression key.
-fn generate_packing_key_switching_key_shares<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    private_glwe_compute_key_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_compression_key: &CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    params: &DistributedCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    preprocessing: &mut P,
-) -> anyhow::Result<LwePackingKeyswitchKeyShares<Z, EXTENSION_DEGREE>>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let noise_vec = preprocessing
-        .next_noise_vec(params.ksk_num_noise, params.ksk_noisebound)?
-        .iter()
-        .map(|share| share.value())
-        .collect_vec();
-
-    mpc_encryption_rng.fill_noise(noise_vec);
-
-    Ok(allocate_and_generate_lwe_packing_keyswitch_key(
-        private_glwe_compute_key_as_lwe,
-        &private_compression_key.post_packing_ks_key,
-        params.raw_compression_parameters.packing_ks_base_log,
-        params.raw_compression_parameters.packing_ks_level,
-        EncryptionType::Bits64,
-        mpc_encryption_rng,
-    ))
-}
-
-#[instrument(name="Gen Compression and Decompression Key", skip(private_glwe_compute_key_as_lwe, private_glwe_compute_key, private_compression_key, mpc_encryption_rng, session, preprocessing), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compression_decompression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    private_glwe_compute_key_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_glwe_compute_key: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_compression_key: &CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    params: DistributedCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-) -> anyhow::Result<(CompressionKey, DecompressionKey)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let packing_key_switching_key_shares = generate_packing_key_switching_key_shares(
-        private_glwe_compute_key_as_lwe,
-        private_compression_key,
-        &params,
-        mpc_encryption_rng,
-        preprocessing,
-    )?;
-
-    let packing_key_switching_key = packing_key_switching_key_shares
-        .open_to_tfhers_type::<u64, _>(session)
-        .await?;
-
-    let compression_key = CompressionKey {
-        packing_key_switching_key,
-        lwe_per_glwe: params.raw_compression_parameters.lwe_per_glwe,
-        storage_log_modulus: params.raw_compression_parameters.storage_log_modulus,
-    };
-
-    let decompression_key = generate_decompression_keys(
-        private_glwe_compute_key,
-        private_compression_key,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-    )
-    .await?;
-
-    Ok((compression_key, decompression_key))
-}
-
-/// Note that in the return value, it is possible for [CompressionPrivateKeyShares] to be None
-/// while [CompressionKey], [DecompressionKey] exists because we do not copy the secret shares again.
-async fn distributed_keygen_compression_material<
-    Z: BaseRing,
-    S: BaseSessionHandles,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + Send + ?Sized,
-    const EXTENSION_DEGREE: usize,
->(
-    session: &mut S,
-    preprocessing: &mut P,
-    params: DKGParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<
-        Z,
-        SoftwareRandomGenerator,
-        EXTENSION_DEGREE,
-    >,
-    glwe_sk_share_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    glwe_secret_key_share: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    glwe_secret_key_share_compression: Option<&CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>>,
-) -> anyhow::Result<(
-    Option<CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>>,
-    Option<(CompressionKey, DecompressionKey)>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let params_basics_handle = params.get_params_basics_handle();
-
-    let compression_material = if let Some(comp_params) =
-        params_basics_handle.get_compression_decompression_params()
-    {
-        match glwe_secret_key_share_compression {
-            Some(inner) => {
-                let compression_keys = generate_compression_decompression_keys(
-                    glwe_sk_share_as_lwe,
-                    glwe_secret_key_share,
-                    inner,
-                    comp_params,
-                    mpc_encryption_rng,
-                    session,
-                    preprocessing,
-                )
-                .await?;
-                (None, Some(compression_keys))
-            }
-            None => {
-                let private_compression_key = CompressionPrivateKeyShares::new_from_preprocessing(
-                    comp_params.raw_compression_parameters,
-                    preprocessing,
-                )?;
-
-                let compression_keys = generate_compression_decompression_keys(
-                    glwe_sk_share_as_lwe,
-                    glwe_secret_key_share,
-                    &private_compression_key,
-                    comp_params,
-                    mpc_encryption_rng,
-                    session,
-                    preprocessing,
-                )
-                .await?;
-                (Some(private_compression_key), Some(compression_keys))
-            }
-        }
-    } else {
-        (None, None)
-    };
-    Ok(compression_material)
-}
-
-#[allow(clippy::too_many_arguments)]
-#[instrument(name="Gen compressed Compression and Decompression Key", skip(private_glwe_compute_key_as_lwe, private_glwe_compute_key, private_compression_key, mpc_encryption_rng, session, preprocessing, seed), fields(sid = ?session.session_id(), own_identity = ?session.own_identity()))]
-async fn generate_compressed_compression_decompression_keys<
-    Z: BaseRing,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
-    S: BaseSessionHandles,
-    Gen: ByteRandomGenerator,
-    const EXTENSION_DEGREE: usize,
->(
-    private_glwe_compute_key_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_glwe_compute_key: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    private_compression_key: &CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>,
-    params: DistributedCompressionParameters,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<Z, Gen, EXTENSION_DEGREE>,
-    session: &mut S,
-    preprocessing: &mut P,
-    seed: u128,
-) -> anyhow::Result<(CompressedCompressionKey, CompressedDecompressionKey)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let packing_key_switching_key_shares = generate_packing_key_switching_key_shares(
-        private_glwe_compute_key_as_lwe,
-        private_compression_key,
-        &params,
-        mpc_encryption_rng,
-        preprocessing,
-    )?;
-
-    let packing_key_switching_key = packing_key_switching_key_shares
-        .open_to_tfhers_seeded_type::<u64, _>(seed, session)
-        .await?;
-
-    let compression_key = CompressedCompressionKey {
-        packing_key_switching_key,
-        lwe_per_glwe: params.raw_compression_parameters.lwe_per_glwe,
-        storage_log_modulus: params.raw_compression_parameters.storage_log_modulus,
-    };
-
-    let decompression_key = generate_compressed_decompression_keys(
-        private_glwe_compute_key,
-        private_compression_key,
-        params,
-        mpc_encryption_rng,
-        session,
-        preprocessing,
-        seed,
-    )
-    .await?;
-
-    Ok((compression_key, decompression_key))
-}
-
-#[allow(clippy::too_many_arguments)]
-/// Note that in the return value, it is possible for [CompressionPrivateKeyShares] to be None
-/// while [CompressionKey], [DecompressionKey] exists because we do not copy the secret shares again.
-async fn distributed_keygen_compressed_compression_material<
-    Z: BaseRing,
-    S: BaseSessionHandles,
-    P: DKGPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + Send + ?Sized,
-    const EXTENSION_DEGREE: usize,
->(
-    session: &mut S,
-    preprocessing: &mut P,
-    params: DKGParams,
-    mpc_encryption_rng: &mut MPCEncryptionRandomGenerator<
-        Z,
-        SoftwareRandomGenerator,
-        EXTENSION_DEGREE,
-    >,
-    glwe_sk_share_as_lwe: &LweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    glwe_secret_key_share: &GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    glwe_secret_key_share_compression: Option<&CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>>,
-    seed: u128,
-) -> anyhow::Result<(
-    Option<CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>>,
-    Option<(CompressedCompressionKey, CompressedDecompressionKey)>,
-)>
-where
-    ResiduePoly<Z, EXTENSION_DEGREE>: ErrorCorrect,
-{
-    let params_basics_handle = params.get_params_basics_handle();
-
-    let compression_material = if let Some(comp_params) =
-        params_basics_handle.get_compression_decompression_params()
-    {
-        match glwe_secret_key_share_compression {
-            Some(inner) => {
-                let compression_keys = generate_compressed_compression_decompression_keys(
-                    glwe_sk_share_as_lwe,
-                    glwe_secret_key_share,
-                    inner,
-                    comp_params,
-                    mpc_encryption_rng,
-                    session,
-                    preprocessing,
-                    seed,
-                )
-                .await?;
-                (None, Some(compression_keys))
-            }
-            None => {
-                let private_compression_key = CompressionPrivateKeyShares::new_from_preprocessing(
-                    comp_params.raw_compression_parameters,
-                    preprocessing,
-                )?;
-
-                let compression_keys = generate_compressed_compression_decompression_keys(
-                    glwe_sk_share_as_lwe,
-                    glwe_secret_key_share,
-                    &private_compression_key,
-                    comp_params,
-                    mpc_encryption_rng,
-                    session,
-                    preprocessing,
-                    seed,
-                )
-                .await?;
-                (Some(private_compression_key), Some(compression_keys))
-            }
-        }
-    } else {
-        (None, None)
-    };
-    Ok(compression_material)
 }
 
 trait Finalizable<const EXTENSION_DEGREE: usize> {
@@ -1916,7 +960,7 @@ where
 
 /// Correctly decompresses a CompressedDecompressionKey standalone generated from the XOF
 /// and outputs a the raw elements of a [`DecompressionKey`] __BEFORE__ the fourrier transform.
-pub fn decompressed_compressed_standalone_decompression_key_from_xof(
+pub fn decompress_compressed_standalone_decompression_key_from_xof(
     compressed_decompression_key: CompressedDecompressionKey,
 ) -> (LweBootstrapKey<Vec<u64>>, LweCiphertextCount) {
     let CompressedDecompressionKey {
@@ -2100,33 +1144,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = OLD_PARAMS_P32_REAL_WITH_SNS;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_switch_and_squash(
-            temp_dir.path(),
-            params.try_into().unwrap(),
-            num_parties,
-            threshold,
-        );
-
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsSnS>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            false,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            false,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            OLD_PARAMS_P32_REAL_WITH_SNS,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: true,
+                run_shortint_with_compact: false,
+                run_fheuint: true,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2156,26 +1183,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P32_NO_SNS_FGLWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            false,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P32_NO_SNS_FGLWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: false,
+                run_shortint_with_compact: true,
+                run_fheuint: true,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2205,20 +1222,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P8_NO_SNS_FGLWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        //This parameter set isnt big enough to run the fheuint tests
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P8_NO_SNS_FGLWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: false,
+                run_shortint_with_compact: true,
+                run_fheuint: false,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2248,26 +1261,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P32_NO_SNS_LWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            false,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P32_NO_SNS_LWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: false,
+                run_shortint_with_compact: true,
+                run_fheuint: true,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2297,20 +1300,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P8_NO_SNS_LWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        //This parameter set isnt big enough to run the fheuint tests
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P8_NO_SNS_LWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: false,
+                run_shortint_with_compact: true,
+                run_fheuint: false,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2339,33 +1338,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = PARAMS_TEST_BK_SNS;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_switch_and_squash(
-            temp_dir.path(),
-            params.try_into().unwrap(),
-            num_parties,
-            threshold,
-        );
-
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsSnS>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            PARAMS_TEST_BK_SNS,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: true,
+                run_shortint_with_compact: true,
+                run_fheuint: true,
+                run_fheuint_with_compression: true,
+            },
         );
     }
 
@@ -2463,33 +1445,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P32_SNS_FGLWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_switch_and_squash(
-            temp_dir.path(),
-            params.try_into().unwrap(),
-            num_parties,
-            threshold,
-        );
-
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsSnS>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            false,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P32_SNS_FGLWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: true,
+                run_shortint_with_compact: true,
+                run_fheuint: true,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2519,27 +1484,16 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = NIST_PARAMS_P8_SNS_FGLWE;
-        let num_parties = 5;
-        let threshold = 1;
-        let temp_dir = tempfile::tempdir().unwrap();
-
-        run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
-
-        run_switch_and_squash(
-            temp_dir.path(),
-            params.try_into().unwrap(),
-            num_parties,
-            threshold,
-        );
-
-        //This parameter set isnt big enough to run the fheuint tests
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsSnS>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
+        run_keygen_test::<EXTENSION_DEGREE>(
+            NIST_PARAMS_P8_SNS_FGLWE,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: true,
+                run_shortint_with_compact: true,
+                run_fheuint: false,
+                run_fheuint_with_compression: false,
+            },
         );
     }
 
@@ -2569,27 +1523,79 @@ pub mod tests {
         ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
         ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
     {
-        let params = BC_PARAMS_NO_SNS;
-        let num_parties = 5;
-        let threshold = 1;
+        run_keygen_test::<EXTENSION_DEGREE>(
+            BC_PARAMS_NO_SNS,
+            5,
+            1,
+            KeygenTestConfig {
+                run_switch_and_squash: false,
+                run_shortint_with_compact: true,
+                run_fheuint: true,
+                run_fheuint_with_compression: true,
+            },
+        );
+    }
+
+    struct KeygenTestConfig {
+        run_switch_and_squash: bool,
+        run_shortint_with_compact: bool,
+        run_fheuint: bool,
+        run_fheuint_with_compression: bool,
+    }
+
+    fn run_keygen_test<const EXTENSION_DEGREE: usize>(
+        params: DKGParams,
+        num_parties: usize,
+        threshold: usize,
+        config: KeygenTestConfig,
+    ) where
+        ResiduePoly<Z64, EXTENSION_DEGREE>: ErrorCorrect,
+        ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Solve + Invert,
+    {
         let temp_dir = tempfile::tempdir().unwrap();
 
         run_dkg_and_save(params, num_parties, threshold, temp_dir.path(), false);
 
-        run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
-        run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
-            temp_dir.path(),
-            params,
-            num_parties,
-            threshold,
-            true,
-        );
+        if config.run_switch_and_squash {
+            run_switch_and_squash(
+                temp_dir.path(),
+                params.try_into().unwrap(),
+                num_parties,
+                threshold,
+            );
+        }
+
+        match params {
+            DKGParams::WithSnS(_) => {
+                run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsSnS>(
+                    temp_dir.path(),
+                    params,
+                    num_parties,
+                    threshold,
+                    config.run_shortint_with_compact,
+                );
+            }
+            DKGParams::WithoutSnS(_) => {
+                run_tfhe_computation_shortint::<EXTENSION_DEGREE, DKGParamsRegular>(
+                    temp_dir.path(),
+                    params,
+                    num_parties,
+                    threshold,
+                    config.run_shortint_with_compact,
+                );
+            }
+        }
+
+        // Only run fheuint tests for parameter sets that are large enough
+        if config.run_fheuint {
+            run_tfhe_computation_fheuint::<EXTENSION_DEGREE>(
+                temp_dir.path(),
+                params,
+                num_parties,
+                threshold,
+                config.run_fheuint_with_compression,
+            );
+        }
     }
 
     // taken from https://stackoverflow.com/questions/64498617/how-to-transpose-a-vector-of-vectors-in-rust
@@ -2609,8 +1615,8 @@ pub mod tests {
     }
 
     #[cfg(feature = "slow_tests")]
-    fn binary_vec64_to_shares128<R: rand::Rng + rand::CryptoRng, const EXTENSION_DEGREE: usize>(
-        v: Vec<u64>,
+    fn binary_vec_to_shares128<R: rand::Rng + rand::CryptoRng, const EXTENSION_DEGREE: usize>(
+        v: Vec<impl Into<u128>>,
         n: usize,
         t: usize,
         rng: &mut R,
@@ -2623,31 +1629,7 @@ pub mod tests {
         transpose2(
             v.into_iter()
                 .map(|s| {
-                    assert!(s == 0 || s == 1);
-                    let s = ResiduePoly::<_, EXTENSION_DEGREE>::from_scalar(Wrapping::<u128>(
-                        s as u128,
-                    ));
-                    ShamirSharings::share(rng, s, n, t).unwrap().shares
-                })
-                .collect::<Vec<Vec<_>>>(),
-        )
-    }
-
-    #[cfg(feature = "slow_tests")]
-    fn binary_vec128_to_shares128<R: rand::Rng + rand::CryptoRng, const EXTENSION_DEGREE: usize>(
-        v: Vec<u128>,
-        n: usize,
-        t: usize,
-        rng: &mut R,
-    ) -> Vec<Vec<crate::execution::sharing::share::Share<ResiduePoly<Z128, EXTENSION_DEGREE>>>>
-    where
-        ResiduePoly<Z128, EXTENSION_DEGREE>: ErrorCorrect + Invert + Solve,
-    {
-        use crate::execution::sharing::shamir::{InputOp, ShamirSharings};
-        use std::num::Wrapping;
-        transpose2(
-            v.into_iter()
-                .map(|s| {
+                    let s: u128 = s.into();
                     assert!(s == 0 || s == 1);
                     let s = ResiduePoly::<_, EXTENSION_DEGREE>::from_scalar(Wrapping::<u128>(s));
                     ShamirSharings::share(rng, s, n, t).unwrap().shares
@@ -2697,10 +1679,10 @@ pub mod tests {
 
         // and then secret share the secret keys
         let compression_key_shares_1 =
-            binary_vec64_to_shares128(compression_key_1, num_parties, threshold, &mut rng);
+            binary_vec_to_shares128(compression_key_1, num_parties, threshold, &mut rng);
         assert_eq!(compression_key_shares_1.len(), num_parties);
         let glwe_key_shares_2 =
-            binary_vec64_to_shares128(glwe_key_2, num_parties, threshold, &mut rng);
+            binary_vec_to_shares128(glwe_key_2, num_parties, threshold, &mut rng);
         assert_eq!(glwe_key_shares_2.len(), num_parties);
 
         // We need to pass these shares into the FnMut,
@@ -2877,7 +1859,7 @@ pub mod tests {
 
         // and then secret share the secret keys
         let glwe_sns_key_shares =
-            binary_vec128_to_shares128(glwe_sns_key, num_parties, threshold, &mut rng);
+            binary_vec_to_shares128(glwe_sns_key, num_parties, threshold, &mut rng);
         assert_eq!(glwe_sns_key_shares.len(), num_parties);
 
         // We need to pass these shares into the FnMut,
