+++ core/threshold/src/execution/tfhe_internals/sns_compression_key.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use tfhe::{
     boolean::prelude::{GlweDimension, PolynomialSize},
-    shortint::parameters::CompressionParameters,
+    shortint::parameters::NoiseSquashingCompressionParameters,
     Versionize,
 };
 use tfhe_versionable::VersionsDispatch;
@@ -17,27 +17,25 @@ use crate::{
 use super::{glwe_key::GlweSecretKeyShare, lwe_key::LweSecretKeyShare};
 
 #[derive(Clone, Serialize, Deserialize, VersionsDispatch)]
-pub enum CompressionPrivateKeySharesVersioned<Z: Clone, const EXTENSION_DEGREE: usize> {
-    V0(CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>),
+pub enum SnsCompressionPrivateKeySharesVersioned<Z: Clone, const EXTENSION_DEGREE: usize> {
+    V0(SnsCompressionPrivateKeyShares<Z, EXTENSION_DEGREE>),
 }
 
-///Structure that holds a share of the LWE key
-/// - data contains shares of the key components
 #[derive(Clone, Debug, Serialize, Deserialize, Versionize, PartialEq)]
-#[versionize(CompressionPrivateKeySharesVersioned)]
-pub struct CompressionPrivateKeyShares<Z: Clone, const EXTENSION_DEGREE: usize> {
+#[versionize(SnsCompressionPrivateKeySharesVersioned)]
+pub struct SnsCompressionPrivateKeyShares<Z: Clone, const EXTENSION_DEGREE: usize> {
     pub post_packing_ks_key: GlweSecretKeyShare<Z, EXTENSION_DEGREE>,
-    pub params: CompressionParameters,
+    pub params: NoiseSquashingCompressionParameters,
 }
 
-impl<Z: BaseRing, const EXTENSION_DEGREE: usize> CompressionPrivateKeyShares<Z, EXTENSION_DEGREE>
+impl<Z: BaseRing, const EXTENSION_DEGREE: usize> SnsCompressionPrivateKeyShares<Z, EXTENSION_DEGREE>
 where
     ResiduePoly<Z, EXTENSION_DEGREE>: Ring,
 {
     pub fn new_from_preprocessing<
         P: BitPreprocessing<ResiduePoly<Z, EXTENSION_DEGREE>> + ?Sized,
     >(
-        params: CompressionParameters,
+        params: NoiseSquashingCompressionParameters,
         preprocessing: &mut P,
     ) -> anyhow::Result<Self> {
         let total_size = params.packing_ks_glwe_dimension.0 * params.packing_ks_polynomial_size.0;
